pyproject.toml
[project]
name = "erav4-s2-uploader"
version = "1.0.0"
description = "Animal selector and file upload application with text extraction"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "PyPDF2>=3.0.0",
    "python-docx>=1.1.0",
]

[tool.setuptools]
packages = []  # Don't treat any directories as packages

[tool.setuptools.package-dir]
# Empty - we're not building packages